Split music and TTS commands into individual modules
